requirements.txt
open3d
tabulate
pillow
matplotlib
opencv-python
flask
torch
numpy